try parallel
